Scene04 test success

+++ include/matrix.h
@@ -9,14 +9,14 @@
 class Mat4
 {
 public:
-  float m[4][4];
+  double m[4][4];
 
   Mat4();
-  Mat4(const float mat[4][4]);
-  Mat4(float m00, float m01, float m02, float m03,
-       float m10, float m11, float m12, float m13,
-       float m20, float m21, float m22, float m23,
-       float m30, float m31, float m32, float m33);
+  Mat4(const double mat[4][4]);
+  Mat4(double m00, double m01, double m02, double m03,
+       double m10, double m11, double m12, double m13,
+       double m20, double m21, double m22, double m23,
+       double m30, double m31, double m32, double m33);
 
   Mat4 operator=(const Mat4 &m2);
   Mat4 operator+(const Mat4 &m2) const;
